Refactor(api and client) se hace validacion de contraseña en el registro de usuario y al iniciar el servidor se carga un usuario como administrador, se arreglan carpetas en components
