+++ api/src/models/User.js
@@ -25,7 +25,7 @@ module.exports = (sequelize) => {
       },
       is_admin: {
         type: DataTypes.BOOLEAN,
-        allowNull: false,
+        defaultValue: false
       },
       e_mail: {
         type: DataTypes.STRING,
