+++ client/src/App.js
@@ -10,8 +10,8 @@ import { FormSponsor } from "./components/FormSponsor/FormSponsor.jsx";
 import { FormUser } from "./components/FormUser/FormUser.jsx";
 import { FormTournament } from "./components/Tournament/FormTournament.jsx";
 import  LogIn  from "./components/LogIn/LogIn.jsx";
-import {FormUserRegister} from "./components/FormUser copy/FormUserRegister.jsx"
-import ControlPanel from "./controlPanel/ControlPanel.jsx";
+import {FormUserRegister} from "./components/FormUserRegister/FormUserRegister.jsx"
+import ControlPanel from "./components/controlPanel/ControlPanel.jsx";
 
 function App() {
   return (
