+++ client/src/components/controlPanel/ControlPanel.jsx
@@ -1,10 +1,11 @@
-import { getAllSponsors, getAllUsers } from "../redux/actions";
+import { getAllSponsors, getAllUsers } from "../../redux/actions";
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import ControlCardUsers from "./ControlCardUsers";
 
 import ControlCardSponsor from "./ControlCardSponsor";
+
 
 export default function ControlPanel() {
   const users = useSelector((state) => state.users);
